sagas/return-order-saga/requirements.txt
flask>=3.1.3
flask-cors>=6.0.2
gunicorn>=25.1.0
kafka-python>=2.3.0
requests>=2.32.0
